+++ network/nm_setting_vpn_pptp_autogen.go
@@ -51,7 +51,7 @@ func getSettingVpnPptpDefaultValue(key string) (value interface{}) {
 	case NM_SETTING_VPN_PPTP_KEY_USER:
 		value = ""
 	case NM_SETTING_VPN_PPTP_KEY_PASSWORD_FLAGS:
-		value = 0
+		value = uint32(0)
 	case NM_SETTING_VPN_PPTP_KEY_PASSWORD:
 		value = ""
 	case NM_SETTING_VPN_PPTP_KEY_DOMAIN:
